Deleting wrt listName of custom routes completed

+++ app.js
@@ -107,14 +107,24 @@ app.post("/", (req, res) => {
 
 app.post("/delete", (req, res) => {
 	const checkedItemId = req.body.checkbox
-	Item.findByIdAndDelete(checkedItemId, (err) => {
-		if (err) {
-			console.log(err)
-		} else {
-			console.log(`Successfully deleted item with id ${checkedItemId}`)
-			res.redirect("/")
-		}
-	})
+	const listName = req.body.listName
+
+	if (listName === "Today") {
+		Item.findByIdAndDelete(checkedItemId, (err) => {
+			if (err) {
+				console.log(err)
+			} else {
+				console.log(`Successfully deleted item with id ${checkedItemId}`)
+				res.redirect("/")
+			}
+		})
+	} else {
+		List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: checkedItemId } } }, (err, foundList) => {
+			if (!err) {
+				res.redirect("/" + listName)
+			}
+		})
+	}
 })
 
 app.listen("4000", () => {
